requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
pydantic==2.6.1
openai==1.12.0
python-dotenv==1.0.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
pillow>=10.0.0